move `SearchCache` to package `*.search.cache`
